requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
flask==2.3.3
requests==2.31.0
gunicorn==21.2.0